// services/flight-service/src/flights/dto/create-flight.dto.d.ts
export declare class CreateFlightDto {
    flightNumber: string;
    origin: string;
    destination: string;
    scheduledDep: string;
    scheduledArr: string;
    aircraftId: string;
    availableSeats: number;
    gate?: string;
    terminal?: string;
}
//# sourceMappingURL=create-flight.dto.d.ts.map